GH-186 moving the `credentialId` back to the `Workspace` where it belongs
