+++ core/src/test/resources/workspace.oauth.properties
@@ -18,6 +18,8 @@ aio_consumer_org_id=aio_consumer_org_id_changeMe
 aio_ims_org_id=aio_ims_org_id_changeMe
 # aio_workspace_id = your Adobe Developer Console workspace Id (project.workspace.id)
 aio_workspace_id=aio_workspace_id_changeMe
+# aio_credential_id = your Adobe Developer Console credential id (project.workspace.details.credentials[i].id)
+aio_credential_id=aio_credential_id_changeMe
 
 # aio_api_key = your Adobe Developer Console credential API Key (or Client ID) (project.workspace.details.credentials[i].oauth_server_to_server.client_id)
 aio_api_key=aio_api_key_changeMe
