+++ ims/src/test/resources/workspace.jwt.properties
@@ -18,7 +18,7 @@ aio_consumer_org_id=aio_consumer_org_id_changeMe
 aio_ims_org_id=aio_ims_org_id_changeMe
 # aio_workspace_id = your Adobe Developer Console workspace Id (project.workspace.id)
 aio_workspace_id=aio_workspace_id_changeMe
-# aio_credential_id = your Adobe Developer Console jwt credential id (project.workspace.details.credentials[i].id)
+# aio_credential_id = your Adobe Developer Console credential id (project.workspace.details.credentials[i].id)
 aio_credential_id=aio_credential_id_changeMe
 # aio_client_secret = your Adobe Developer Console jwt credential client secret (project.workspace.details.credentials[i].jwt.client_secret)
 aio_client_secret=aio_client_secret_changeMe
